Inserted retry logic when Jetty throws ConcurrentModificationException

+++ common/s2i/init-db-from-war.sh
@@ -72,20 +72,29 @@ export JETTY_PID=$(echo $!)
 echo "JETTY_PID=${JETTY_PID}"
 sleep 3
 tail -f db_creation.log &
-for i in {1..360}
+for i in {1..720}
 do
     sleep 1
     if [[ -f db_creation.log ]] &&  fgrep --quiet "oejs.Server:main: Started" "db_creation.log" ; then
-        if [[ -d "./protected/databaseBackups/develop/portDataSource" ]]; then
-          echo "An Entando backup was detected. Please verify that your data has been restored."
-        fi
-        echo "Jetty started and database created."
-# Attempt killing Jetty only AFTER waiting for it to terminate
+    # Attempt killing Jetty only AFTER waiting for it to terminate
         (echo "Waiting for Jetty process [$JETTY_PID] to shut down"; sleep 3; ps; kill -9 ${JETTY_PID}; ps) &
         wait ${JETTY_PID}
-        popd
-        rm -Rf /tmp/entando-db-build
-        exit 0
+        if fgrep --quiet "java.util.ConcurrentModificationException" "db_creation.log" ; then
+            mv db_creation.log db_creation_conccurency_exception.log
+            $jetty_command  &> db_creation.log &
+            export JETTY_PID=$(echo $!)
+            echo "Restarting Jetty due to ConcurrentModificationException. JETTY_PID=${JETTY_PID}"
+            sleep 3
+            tail -f db_creation.log &
+        else
+            if [[ -d "./protected/databaseBackups/develop/portDataSource" ]]; then
+              echo "An Entando backup was detected. Please verify that your data has been restored."
+            fi
+            echo "Jetty started and database created."
+            popd
+            rm -Rf /tmp/entando-db-build
+            exit 0
+        fi
     fi
 done
 echo "BUILD TIMED OUT!!!!!"
